pyproject.toml
[project]
name = "hackathon-multiagent-system"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.14,<4.0"
dependencies = [
    "langgraph (>=1.1.3,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "langchain (>=1.2.13,<2.0.0)",
    "langchain-openai (>=1.1.11,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "black (>=26.3.1,<27.0.0)",
    "rdkit (>=2025.9.6,<2026.0.0)",
    "requests (>=2.32.0,<3.0.0)",
    "beautifulsoup4 (>=4.12.0,<5.0.0)",
    "markdownify (>=0.14.1,<0.15.0)",
    "scikit-learn (>=1.5.0,<2.0.0)",
    "numpy (>=2.1.0,<3.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
